--- src/main/java/com/quantumzone/QZ_Workhub/domain/dto/NotificacionDto.java
@@ -1,5 +1,0 @@
-package com.quantumzone.QZ_Workhub.domain.dto;
-
-public class NotificacionDto {
-
-}
