--- src/main/java/com/quantumzone/QZ_Workhub/domain/dto/ReservaDto.java
@@ -1,4 +1,0 @@
-package com.quantumzone.QZ_Workhub.domain.dto;
-
-public class ReservaDto {
-}
