--- src/main/java/com/quantumzone/QZ_Workhub/domain/dto/SalaDto.java
@@ -1,4 +1,0 @@
-package com.quantumzone.QZ_Workhub.domain.dto;
-
-public class SalaDto {
-}
