--- src/main/java/com/quantumzone/QZ_Workhub/domain/dto/UsuarioDto.java
@@ -1,4 +1,0 @@
-package com.quantumzone.QZ_Workhub.domain.dto;
-
-public class UsuarioDto {
-}
