+++ src/main/java/com/quantumzone/QZ_Workhub/dominio/dto/NotificacionDto.java
@@ -1,0 +1,5 @@
+package com.quantumzone.QZ_Workhub.dominio.dto;
+
+public class NotificacionDto {
+
+}
