+++ src/main/java/com/quantumzone/QZ_Workhub/dominio/dto/PagoDto.java
@@ -1,0 +1,4 @@
+package com.quantumzone.QZ_Workhub.dominio.dto;
+
+public class PagoDto {
+}
