+++ src/main/java/com/quantumzone/QZ_Workhub/dominio/dto/RecursoReservado.java
@@ -1,0 +1,4 @@
+package com.quantumzone.QZ_Workhub.dominio.dto;
+
+public class RecursoReservado {
+}
